% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateR_NCBI.R
\name{annotateR_NCBI}
\alias{annotateR_NCBI}
\title{annotateR_NCBI}
\usage{
annotateR_NCBI(a, b)
}
\arguments{
\item{a}{NCBI BED file containing 'XR/XM' nomenclature for genes.}

\item{b}{NCBI 'all fields' file containing the alternate gene names.}
}
\value{
Returns an annotated BED file containing both gene names.
}
\description{
Changes the gene names of an NCBI BED file.
}
\examples{
\dontrun{}

}
