% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeds.R
\docType{data}
\name{breeds}
\alias{breeds}
\title{Dataset containing breeds}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{Breed}{Names of breeds}
}
}
\usage{
data(breeds)
}
\description{
Example dataset containing several breeds referring
to the samples from the example
SNPs.recode.subset.rename.vcf.gz file
located in the inst/extdata folder.
}
\examples{

data(breeds)
head(breeds)
}
\references{
This data set was created for the variantscanR package as an example for the diveRsity function.
}
\keyword{datasets}
