% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomenameR.R
\name{chromosomenameR}
\alias{chromosomenameR}
\title{chromosomenameR}
\usage{
chromosomenameR(a, b)
}
\arguments{
\item{a}{VCF file that needs a different chromosome nomenclature}

\item{b}{Conversion file that contains the original nomenclature in the first column and the desired nomenclature in the second.}
}
\value{
Returns a VCF file with different chromosome nomenclature
}
\description{
Changes 'NC_xxx' chromosome nomenclature to 'Chr' nomenclature.
}
\examples{
\donttest{}
}
