% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diveRsity.R
\name{diveRsity}
\alias{diveRsity}
\title{diveRsity}
\usage{
diveRsity(VCF, breeds, sample_name)
}
\arguments{
\item{VCF}{Multisample VCF file}

\item{breeds}{An Excel file containing the breed of every sample.}

\item{sample_name}{A string containing the name of the sample of interest that needs to be analysed.}
}
\value{
A dataframe containing the level of heterozygosity of every sample, together with a graph, representing the same information.
}
\description{
Provides a measure of diversity based on the percentage of heterozygosity per sample.
}
\examples{
\dontrun{}
}
