% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_chromosome_names.R
\docType{data}
\name{dog_chromosome_names}
\alias{dog_chromosome_names}
\title{File containing the names of the chromosomes of the domestic dog (Canis familiaris)}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{RefSeq.Accn}{RefSeq id names for Canis familiaris chromosomes}
 \item{UCSC style name}{UCSC style names for Canis familiaris chromosome}
}
}
\usage{
data(dog_chromosome_names)
}
\description{
File containing the names of the chromosomes of the domestic dog (Canis familiaris)
}
\examples{

data(dog_chromosome_names)
head(dog_chromosome_names)
}
\references{
This data manually created for the variantscanR package
}
\keyword{datasets}
