% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafiltR.R
\name{extrafiltR}
\alias{extrafiltR}
\title{extrafiltR}
\usage{
extrafiltR(vcf, variants_file, BED_file_annot)
}
\arguments{
\item{vcf}{Single sample VCF file, uploaded with the vcfscanneR function.}

\item{variants_file}{File containing variants of interest, processed by the pRocess function.}

\item{BED_file_annot}{Annotated BED file created with the annotateR_NCBI/ensembl function.}
}
\value{
A dataframe will be obtained containing all the variants within the genes of interest.
}
\description{
Filtering function that takes the single sample VCF file and filters for present variants in genes of
interest. These genes of interest are obtained from the user-defined variants file.
}
\examples{
\dontrun{}
}
