% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRocess.R
\name{pRocess}
\alias{pRocess}
\title{pRocess}
\usage{
pRocess(
  variants_file,
  BED_file,
  refseq,
  QC = TRUE,
  organism = NULL,
  OMIA = FALSE
)
}
\arguments{
\item{variants_file}{File containing variants of interest. Important: Specific format needed, check vignette for more information}

\item{BED_file}{Annotated BED file processed by annotateR_NCBI() or annotateR_ensembl() function}

\item{refseq}{Reference genome assembly (OPTIONAL) e.g. Canfam3.1}

\item{QC}{Quality Control (OPTIONAL). If TRUE (default) the function checks whether the nucleotide in the variants_file is the same as that of the reference genome}

\item{organism}{Optional argument but mandatory if QC = TRUE. This should be a standard BSgenome organism code (list with available organisms and codes is provided in the DATA folder)}

\item{OMIA}{Optional argument which is by default FALSE. If TRUE, the variants_file will be handled differently. OMIA should be TRUE if the organism file is directly downloaded from the OMIA website.}
}
\value{
Returns a processed file containing the variants of interest, ready for downstream analysis.
}
\description{
Optional function that processes the user-defined file with the variants of interest. If desired, this function performes some quality control steps. This quality control might include providing a reference genome assembly (refseq argument, e.g. CanFam3.1) and removes all variants from other assemblies.
}
\examples{
\dontrun{}
}
