% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_file_alternative.R
\docType{data}
\name{variants_file_alternative}
\alias{variants_file_alternative}
\title{File containing the variants of interest in an alternative format.}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{Chromosome}{The name of chromosome}
 \item{Start}{Position start of the variant}
 \item{End}{Position end of the varians (in case of insertion/deletion)}
 \item{Reference}{The name of chromosome}
 \item{Reference Sequence}{Reference sequence chromosome or scaffold}
 \item{Inheritance pattern}{Inheritance pattern}
 \item{Gene}{Gene in which the variant is located}
 \item{Variant Phenotype}{Variant phenotype}
 \item{Breed}{Breeds in which the variant has been found}
 \item{OMIA ID(s)}{OMIA variant ID}
 \item{Species Name}{Name of species in which the variant was described}
 \item{Allele}{Allele}
 \item{Type of Variant}{Type of variant}
 \item{Deleterious?}{Is the individual's susceptibility or predisposition incread to a certain disease or disorder?}
 \item{c. or n.}{Coding or non-codion region}
 \item{p.}{Protein}
 \item{EVA ID}{European Variation Archive ID}
 \item{Year published}{Year in which the variant was first published}
 \item{PubMed ID(s)}{PubMed ID(s) of publication(s) referring to the variant}
}
}
\usage{
data(variants_file_alternative)
}
\description{
File containing the variants of interest for the dog.
This file was manually created for demonstrative purposes.
}
\examples{

data(variants_file_alternative)
head(variants_file_alternative)
}
\references{
This data set was manually created for demonstrative purposes for the variantscanR package.
}
\keyword{datasets}
