% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfscanneR.R
\name{vcfscanneR}
\alias{vcfscanneR}
\title{vcfscanneR}
\usage{
vcfscanneR(x, sample)
}
\arguments{
\item{x}{Character string containing the name of the vcf file or file path.}

\item{sample}{Character string containing the name of the sample of interest. The sample name has to match with that of the VCF file.}
}
\value{
Returns a single sample vcf file ready for downstream analysis.
}
\description{
Uploads a single sample VCF file (from a multi sample VCF file) into R.
}
\examples{
\donttest{}
}
